For atomic logging to work, only a single write call should be issued per log line. To accomplish this, the partial logging functions instead write to a line buffer, and then are flushed out to the streams at the end. 'flushLog' was repurposed for this. Buffer size is controlled by the existing macro 'SentryCrashLOGGER_CBufferSize', which now has its default size increased to account for emoji streams and ligatures for colorized console fonts. setLogFD no longer allows setting the g_fd or g_file to stdout, etc., because stdout was always written anyway, so this part was cleaned up.

+++ Sources/Sentry/NSString+SentryNSUIntegerValue.m
@@ -19,7 +19,7 @@
 @implementation NSString (SentryNSUIntegerValue)
 
 - (NSUInteger)unsignedLongLongValue {
-    return strtoull([self UTF8String], NULL, 0);
+    return (NSUInteger)strtoull([self UTF8String], NULL, 0);
 }
 
 @end
